+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 ### Changed
 
  - The library must now be compiled with C++ 17 compiler
- - If the user sets its custom allocator, the return allocated memory must now be 16 bytes aligned
+ - If the user sets its own custom allocator, the return allocated memory must now be 16 bytes aligned
+ - The internal allocators now allocates memory that is 16-bytes aligned
 
 ### Removed